db file with type affinity
SQLite is not strictly typed, though the first db file had no type affinity.
Created a new db file with type affinity, this results in that a model_year = '2000' will be that the model_year is an integer instead of a string.
This enables better use of comparison searches/etc